Refactor popup to have main module
