+++ package.json
@@ -27,6 +27,7 @@
     "copy-webpack-plugin": "^4.0.1",
     "css-loader": "^0.28.3",
     "eslint": "^4.1.1",
+    "file-loader": "^0.11.2",
     "html-loader": "^0.4.5",
     "html-webpack-plugin": "^2.28.0",
     "rimraf": "^2.6.1",
