+++ src/popup/main/main.css
@@ -1,0 +1,13 @@
+
+.time-label, .char-label {
+	margin: 1px 5px;
+}
+
+.char-label {
+	display: block;
+	transition: color .1s;
+}
+
+.full-char-label {
+	color: #D32F2F;
+}
