+++ src/popup/main/main.html
@@ -1,0 +1,23 @@
+<div>
+    <div class="top-bar">
+        <button class="icon-btn options-btn">
+            <i class="material-icons">settings</i>
+        </button>
+        <div class="comment-field">
+        <input id="input-field"
+                class="input-field"
+                type = 'text'
+                maxlength='140'
+                placeholder="insightful comment"
+                autofocus/>
+        <span class="char-label" id="char-label">0/140</span>
+        </div>
+        <button class="submit-btn">
+            <i class="material-icons">send</i>
+        </button>
+    </div>
+
+    <div class="btn-bar">
+    </div>
+    <button class="text-btn logout-btn">Logout</button>
+</div>
