+++ src/popup/options/options.html
@@ -1,9 +1,7 @@
-<div class="popup-section">
+<div>
     <button class="icon-btn back-btn">
         <i class="material-icons">arrow_back</i>
     </button>
     <button class="text-btn">Option</button>
     <button class="text-btn">Option</button>
-
-
 </div>
