+++ src/popup/popup.html
@@ -8,35 +8,10 @@
 	<body>
 		<div class="popup-body">
 		<div class="popup-inner">
-			<div class="popup-section main-section">
-				<div class="top-bar">
-					<button class="icon-btn options-btn">
-						<i class="material-icons">settings</i>
-					</button>
-					<div class="comment-field">
-					<input id="input-field"
-							class="input-field"
-							type = 'text'
-							maxlength='140'
-							placeholder="insightful comment"
-							autofocus/>
-					<span class="char-label" id="char-label">0/140</span>
-					</div>
-					<button class="submit-btn">
-						<i class="material-icons">send</i>
-					</button>
-				</div>
-				
-
-				<div class="btn-bar">
-				</div>
-				<button class="text-btn logout-btn">Logout</button>
-			</div>
-			<div class="popup-section options-section"></div>
+			<main class="popup-section main-section"></main>
+			<options class="popup-section options-section"></options>
 		</div>
 		</div>
-		
-
 	</body>
 
 </html>
